requirements.txt
langchain==0.1.0
streamlit==1.32.0
python-telegram-bot==20.8
python-dotenv==1.0.0
openai==1.12.0
requests==2.31.0
tiktoken==0.6.0 